debug/barcode_generator/pyproject.toml
[project]
name = "barcode-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "python-barcode>=0.16.1",
]
